advertisements_with_restrictions/requirements.txt
python-dotenv
django
djangorestframework
psycopg2-binary
django-filter
gunicorn